+++ frontend/src/app/main/ui/workspace/sidebar/options.cljs
@@ -12,6 +12,7 @@
    [app.common.files.helpers :as cfh]
    [app.common.geom.shapes :as gsh]
    [app.common.types.shape.layout :as ctl]
+   [app.main.data.helpers :as dsh]
    [app.main.data.workspace :as udw]
    [app.main.data.workspace.common :as dwc]
    [app.main.refs :as refs]
@@ -23,7 +24,6 @@
    [app.main.ui.workspace.sidebar.options.menus.align :refer [align-options*]]
    [app.main.ui.workspace.sidebar.options.menus.bool :refer [bool-options*]]
    [app.main.ui.workspace.sidebar.options.menus.component :refer [component-menu]]
-   [app.main.ui.workspace.sidebar.options.menus.exports :refer [exports-menu]]
    [app.main.ui.workspace.sidebar.options.menus.grid-cell :as grid-cell]
    [app.main.ui.workspace.sidebar.options.menus.interactions :refer [interactions-menu]]
    [app.main.ui.workspace.sidebar.options.menus.layout-container :as layout-container]
@@ -38,13 +38,15 @@
    [app.main.ui.workspace.sidebar.options.shapes.svg-raw :as svg-raw]
    [app.main.ui.workspace.sidebar.options.shapes.text :as text]
    [app.util.i18n :as i18n :refer [tr]]
+   [okulary.core :as l]
+
    [rumext.v2 :as mf]))
 
 ;; --- Options
 
-(mf/defc shape-options*
-  {::mf/wrap [#(mf/throttle % 60)]}
-  [{:keys [shape shapes-with-children page-id file-id libraries] :as props}]
+(mf/defc single-shape-options*
+  {::mf/private true}
+  [{:keys [shape page-id file-id libraries] :as props}]
   (let [shape-type (dm/get-prop shape :type)
         shape-id   (dm/get-prop shape :id)
 
@@ -53,45 +55,83 @@
 
         shape      (gsh/transform-shape shape modifiers)]
 
-    [:*
-     (case shape-type
-       :frame   [:> frame/options* props]
-       :group   [:> group/options* {:shape shape :shape-with-children shapes-with-children :file-id file-id :libraries libraries}]
-       :text    [:> text/options* {:shape shape  :file-id file-id :libraries libraries}]
-       :rect    [:> rect/options* {:shape shape}]
-       :circle  [:> circle/options* {:shape shape}]
-       :path    [:> path/options* {:shape shape}]
-       :svg-raw [:> svg-raw/options* {:shape shape}]
-       :bool    [:> bool/options* {:shape shape}]
-       nil)
-     [:& exports-menu
-      {:ids [(:id shape)]
-       :values (select-keys shape [:exports])
-       :shape shape
-       :page-id page-id
-       :file-id file-id}]]))
+    (case shape-type
+      :frame   [:> frame/options* props]
+      :group   [:> group/options* props]
+      :text    [:> text/options* {:shape shape :file-id file-id :page-id page-id :libraries libraries}]
+      :rect    [:> rect/options* {:shape shape :file-id file-id :page-id page-id}]
+      :circle  [:> circle/options* {:shape shape :file-id file-id :page-id page-id}]
+      :path    [:> path/options* {:shape shape :file-id file-id :page-id page-id}]
+      :svg-raw [:> svg-raw/options* {:shape shape :file-id file-id :page-id page-id}]
+      :bool    [:> bool/options* {:shape shape :file-id file-id :page-id page-id}]
+      nil)))
+
+(mf/defc shape-options*
+  {::mf/wrap [#(mf/throttle % 100)]
+   ::mf/private true}
+  [{:keys [shapes shapes-with-children selected page-id file-id libraries]}]
+  (if (= 1 (count selected))
+    [:> single-shape-options*
+     {:page-id page-id
+      :file-id file-id
+      :libraries libraries
+      :shape (first shapes)
+      :shapes-with-children shapes-with-children}]
+    [:> multiple/options*
+     {:shapes-with-children shapes-with-children
+      :shapes shapes
+      :page-id page-id
+      :file-id file-id
+      :libraries libraries}]))
 
 (mf/defc specialized-panel*
-  {::mf/wrap [mf/memo]}
+  {::mf/private true}
   [{:keys [panel]}]
   (when (= (:type panel) :component-swap)
     [:& component-menu {:shapes (:shapes panel) :swap-opened? true}]))
 
 (mf/defc design-menu*
-  {::mf/wrap [mf/memo]}
-  [{:keys [selected objects page-id file-id selected-shapes shapes-with-children]}]
-  (let [sp-panel             (mf/deref refs/specialized-panel)
-        drawing              (mf/deref refs/workspace-drawing)
-        libraries            (mf/deref refs/libraries)
-        edition              (mf/deref refs/selected-edition)
-        edit-grid?           (ctl/grid-layout? objects edition)
-        grid-edition         (mf/deref refs/workspace-grid-edition)
-        selected-cells       (->> (dm/get-in grid-edition [edition :selected])
-                                  (map #(dm/get-in objects [edition :layout-grid-cells %])))]
+  {::mf/private true}
+  [{:keys [selected objects page-id file-id shapes]}]
+  (let [sp-panel (mf/deref refs/specialized-panel)
+        drawing  (mf/deref refs/workspace-drawing)
+        edition  (mf/deref refs/selected-edition)
+
+        files
+        (mf/deref refs/files)
+
+        libraries
+        (mf/with-memo [files file-id]
+          (refs/select-libraries files file-id))
+
+        edit-grid?
+        (mf/with-memo [objects edition]
+          (ctl/grid-layout? objects edition))
+
+        grid-edition
+        (mf/deref refs/workspace-grid-edition)
+
+        selected-cells
+        (->> (dm/get-in grid-edition [edition :selected])
+             (map #(dm/get-in objects [edition :layout-grid-cells %])))
+
+        shapes-with-children
+        (mf/with-memo [selected objects shapes]
+          (let [xform    (comp (remove nil?)
+                               (mapcat #(cfh/get-children-ids objects %)))
+                selected (into selected xform selected)]
+            (sequence (keep (d/getf objects)) selected)))
+
+
+        total-selected
+        (count selected)]
 
     [:div {:class (stl/css :element-options :design-options)}
-     [:> align-options*]
-     [:> bool-options*]
+     [:> align-options* {:shapes shapes
+                         :objects objects}]
+     [:> bool-options* {:total-selected total-selected
+                        :shapes shapes
+                        :shapes-with-children shapes-with-children}]
 
      (cond
        (and edit-grid? (d/not-empty? selected-cells))
@@ -104,67 +144,71 @@
         {:ids [edition]
          :values (get objects edition)}]
 
-       (not (nil? sp-panel))
+       (some? sp-panel)
        [:> specialized-panel* {:panel sp-panel}]
 
        (d/not-empty? drawing)
        [:> drawing/drawing-options*
         {:drawing-state drawing}]
 
-       (= 0 (count selected))
+       (zero? total-selected)
        [:> page/options*]
 
-       (= 1 (count selected))
-       [:> shape-options*
-        {:shape (first selected-shapes)
-         :page-id page-id
-         :file-id file-id
-         :libraries libraries
-         :shapes-with-children shapes-with-children}]
-
        :else
-       [:> multiple/options*
-        {:shapes-with-children shapes-with-children
-         :shapes selected-shapes
+       [:> shape-options*
+        {:shapes shapes
+         :shapes-with-children shapes-with-children
          :page-id page-id
          :file-id file-id
+         :selected selected
          :libraries libraries}])]))
 
-;; FIXME: need optimizations
+(mf/defc inspect-tab*
+  {::mf/private true}
+  [{:keys [objects shapes] :as props}]
+  (let [frame
+        (cfh/get-frame objects (first shapes))
+
+        props
+        (mf/spread-props props
+                         {:frame frame
+                          :from :workspace})]
+
+    [:> hrs/right-sidebar* props]))
+
+(def ^:private options-tabs
+  [{:label (tr "workspace.options.design")
+    :id "design"}
+   {:label (tr "workspace.options.prototype")
+    :id "prototype"}
+   {:label (tr "workspace.options.inspect")
+    :id "inspect"}])
+
+(defn- on-option-tab-change
+  [mode]
+  (let [mode (keyword mode)]
+    (st/emit! (udw/set-options-mode mode))
+    (if (= mode :inspect)
+      (st/emit! :interrupt (dwc/set-workspace-read-only true))
+      (st/emit! :interrupt (dwc/set-workspace-read-only false)))))
+
 (mf/defc options-content*
-  {::mf/memo true
-   ::mf/private true}
-  [{:keys [selected shapes shapes-with-children page-id file-id on-change-section on-expand]}]
-  (let [objects              (mf/deref refs/workspace-page-objects)
-        permissions          (mf/use-ctx ctx/permissions)
+  {::mf/private true}
+  [{:keys [objects selected page-id file-id on-change-section on-expand]}]
+  (let [permissions
+        (mf/use-ctx ctx/permissions)
 
-        selected-shapes      (into [] (keep (d/getf objects)) selected)
-        first-selected-shape (first selected-shapes)
-        shape-parent-frame   (cfh/get-frame objects (:frame-id first-selected-shape))
+        options-mode
+        (mf/deref refs/options-mode-global)
 
-        options-mode         (mf/deref refs/options-mode-global)
-
-        on-change-tab
-        (fn [options-mode]
-          (let [options-mode (keyword options-mode)]
-            (st/emit! (udw/set-options-mode options-mode))
-            (if (= options-mode :inspect)
-              (st/emit! :interrupt (dwc/set-workspace-read-only true))
-              (st/emit! :interrupt (dwc/set-workspace-read-only false)))))
-
-        tabs
-        (mf/with-memo []
-          [{:label (tr "workspace.options.design")
-            :id "design"}
-           {:label (tr "workspace.options.prototype")
-            :id "prototype"}
-           {:label (tr "workspace.options.inspect")
-            :id "inspect"}])]
+        shapes
+        (mf/with-memo [selected objects]
+          (sequence (keep (d/getf objects)) selected))]
 
     [:div {:class (stl/css :tool-window)}
      (if (:can-edit permissions)
-       [:> tab-switcher* {:tabs tabs
-                          :on-change on-change-tab
+       [:> tab-switcher* {:tabs options-tabs
+                          :on-change on-option-tab-change
                           :selected (name options-mode)
                           :class (stl/css :options-tab-switcher)}
         (case options-mode
@@ -174,46 +218,46 @@
 
           :inspect
           [:div {:class (stl/css :element-options :inspect-options)}
-           [:& hrs/right-sidebar {:page-id           page-id
-                                  :objects           objects
-                                  :file-id           file-id
-                                  :frame             shape-parent-frame
-                                  :shapes            selected-shapes
-                                  :on-change-section on-change-section
-                                  :on-expand         on-expand
-                                  :from              :workspace}]]
+           [:> inspect-tab* {:page-id page-id
+                             :file-id file-id
+                             :objects objects
+                             :selected selected
+                             :shapes shapes
+                             :on-change-section on-change-section
+                             :on-expand on-expand}]]
 
           :design
           [:> design-menu* {:selected selected
                             :objects objects
                             :page-id page-id
                             :file-id file-id
-                            :selected-shapes selected-shapes
-                            :shapes-with-children shapes-with-children}])]
+                            :shapes shapes}])]
 
-       ;; FIXME: Reuse tab???
        [:div {:class (stl/css :element-options :inspect-options :read-only)}
-        [:& hrs/right-sidebar {:page-id           page-id
-                               :objects           objects
-                               :file-id           file-id
-                               :frame             shape-parent-frame
-                               :shapes            selected-shapes
-                               :on-change-section on-change-section
-                               :on-expand         on-expand
-                               :from              :workspace}]])]))
+        [:> inspect-tab* {:page-id page-id
+                          :file-id file-id
+                          :objects objects
+                          :selected selected
+                          :shapes shapes
+                          :on-change-section on-change-section
+                          :on-expand on-expand}]])]))
 
-;; TODO: this need optimizations, selected-objects and
-;; selected-objects-with-children are derefed always but they only
-;; need on multiple selection in majority of cases
+(defn- make-page-objects-ref
+  [file-id page-id]
+  (l/derived #(dsh/lookup-page-objects % file-id page-id) st/state))
 
 (mf/defc options-toolbox*
   {::mf/memo true}
   [{:keys [page-id file-id section selected on-change-section on-expand]}]
-  (let [shapes               (mf/deref refs/selected-objects)
-        shapes-with-children (mf/deref refs/selected-shapes-with-children)]
-    [:> options-content* {:shapes shapes
+  (let [objects-ref
+        (mf/with-memo [page-id file-id]
+          (make-page-objects-ref file-id page-id))
+
+        objects
+        (mf/deref objects-ref)]
+
+    [:> options-content* {:objects objects
                           :selected selected
-                          :shapes-with-children shapes-with-children
                           :file-id file-id
                           :page-id page-id
                           :section section