+++ frontend/src/app/main/ui/workspace/sidebar/options/menus/bool.cljs
@@ -8,11 +8,12 @@
   (:require-macros [app.main.style :as stl])
   (:require
    [app.common.data :as d]
+   [app.common.data.macros :as dm]
+   [app.common.files.helpers :as cfh]
    [app.main.data.workspace.bool :as dwb]
    [app.main.data.workspace.path.shapes-to-path :as dwps]
    [app.main.data.workspace.shortcuts :as sc]
    [app.main.features :as features]
-   [app.main.refs :as refs]
    [app.main.store :as st]
    [app.main.ui.components.radio-buttons :refer [radio-button radio-buttons]]
    [app.main.ui.icons :as i]
@@ -23,48 +24,61 @@
   (i/icon-xref :boolean-flatten (stl/css :flatten-icon)))
 
 (mf/defc bool-options*
-  {::mf/memo true}
-  []
-  (let [selected               (mf/deref refs/selected-objects)
-        head                   (first selected)
-        selected-with-children (mf/deref refs/selected-shapes-with-children)
-        has-invalid-shapes?    (->> selected-with-children
-                                    (some (comp #{:frame :text} :type)))
-        is-group? (and (some? head) (= :group (:type head)))
-        is-bool?  (and (some? head) (= :bool (:type head)))
-        head-bool-type (and (some? head) is-bool? (:bool-type head))
+  [{:keys [total-selected shapes shapes-with-children]}]
+  (let [head      (first shapes)
+        head-id   (dm/get-prop head :id)
 
-        first-not-group-like?
-        (and (= (count selected) 1)
-             (not (contains? #{:group :bool} (:type (first selected)))))
+        is-group? (cfh/group-shape? head)
+        is-bool?  (cfh/bool-shape? head)
+
+        head-bool-type
+        (and is-bool? (get head :bool-type))
+
+        render-wasm-enabled?
+        (features/use-feature "render-wasm/v1")
+
+        has-invalid-shapes?
+        (if render-wasm-enabled?
+          false
+          (some (fn [shape]
+                  (or (cfh/frame-shape? shape)
+                      (cfh/text-shape? shape)))
+                shapes-with-children))
+
+        head-not-group-like?
+        (and (= 1 total-selected)
+             (not is-group?)
+             (not is-bool?))
 
         disabled-bool-btns
-        (if (features/active-feature? @st/state "render-wasm/v1")
+        (if render-wasm-enabled?
           false
-          (or (empty? selected) has-invalid-shapes? first-not-group-like?))
+          (or (zero? total-selected)
+              has-invalid-shapes?
+              head-not-group-like?))
 
         disabled-flatten
-        (if (features/active-feature? @st/state "render-wasm/v1")
+        (if render-wasm-enabled?
           false
-          (or (empty? selected) has-invalid-shapes?))
+          (or (zero? total-selected)
+              has-invalid-shapes?))
 
-        set-bool
+        on-change
         (mf/use-fn
-         (mf/deps  selected is-group?  is-bool?)
+         (mf/deps total-selected is-group? is-bool? head-id head-bool-type)
          (fn [bool-type]
            (let [bool-type (keyword bool-type)]
-
              (cond
-               (> (count selected) 1)
+               (> total-selected 1)
                (st/emit! (dwb/create-bool bool-type))
 
-               (and (= (count selected) 1) is-group?)
-               (st/emit! (dwb/group-to-bool (:id head) bool-type))
+               (and (= total-selected 1) is-group?)
+               (st/emit! (dwb/group-to-bool head-id bool-type))
 
-               (and (= (count selected) 1) is-bool?)
+               (and (= total-selected 1) is-bool?)
                (if (= head-bool-type bool-type)
-                 (st/emit! (dwb/bool-to-group (:id head)))
-                 (st/emit! (dwb/change-bool-type (:id head) bool-type)))))))
+                 (st/emit! (dwb/bool-to-group head-id))
+                 (st/emit! (dwb/change-bool-type head-id bool-type)))))))
 
         flatten-objects
         (mf/use-fn  #(st/emit! (dwps/convert-selected-to-path)))]
@@ -74,7 +88,7 @@
        [:div {:class (stl/css :bool-group)}
         [:& radio-buttons {:selected (d/name head-bool-type)
                            :class (stl/css :boolean-radio-btn)
-                           :on-change set-bool
+                           :on-change on-change
                            :name "bool-options"}
          [:& radio-button {:icon i/boolean-union
                            :value "union"