+++ frontend/src/app/main/ui/workspace/sidebar/options/menus/layout_item.cljs
@@ -309,6 +309,7 @@
                      :title "Align self end"
                      :id    "align-self-end"}]])
 
+
 (mf/defc layout-item-menu
   {::mf/memo #{:ids :values :type :is-layout-child? :is-grid-parent :is-flex-parent? :is-grid-layout? :is-flex-layout?}
    ::mf/props :obj}