+++ frontend/src/app/main/ui/workspace/sidebar/options/shapes/group.cljs
@@ -14,6 +14,7 @@
    [app.main.ui.workspace.sidebar.options.menus.blur :refer [blur-menu]]
    [app.main.ui.workspace.sidebar.options.menus.color-selection :refer [color-selection-menu*]]
    [app.main.ui.workspace.sidebar.options.menus.constraints :refer [constraints-menu]]
+   [app.main.ui.workspace.sidebar.options.menus.exports :refer [exports-menu* exports-attrs]]
    [app.main.ui.workspace.sidebar.options.menus.fill :as fill]
    [app.main.ui.workspace.sidebar.options.menus.grid-cell :as grid-cell]
    [app.main.ui.workspace.sidebar.options.menus.layer :refer [layer-menu]]
@@ -29,7 +30,7 @@
 
 (mf/defc options*
   {::mf/wrap [mf/memo]}
-  [{:keys [shape shapes-with-children libraries file-id] :as props}]
+  [{:keys [shape shapes-with-children libraries file-id page-id]}]
 
   (let [id     (dm/get-prop shape :id)
         type   (dm/get-prop shape :type)
@@ -157,6 +158,14 @@
        [:& ot/text-menu {:type type :ids text-ids :values text-values}])
 
      (when-not (empty? svg-values)
-       [:& svg-attrs-menu {:ids ids :values svg-values}])]))
+       [:& svg-attrs-menu {:ids ids :values svg-values}])
+
+     [:> exports-menu* {:type type
+                        :ids ids
+                        :shapes shapes
+                        :values (select-keys shape exports-attrs)
+                        :page-id page-id
+                        :file-id file-id}]]))
+