+++ frontend/src/app/main/ui/workspace/sidebar/options/shapes/path.cljs
@@ -11,6 +11,7 @@
    [app.main.refs :as refs]
    [app.main.ui.workspace.sidebar.options.menus.blur :refer [blur-menu]]
    [app.main.ui.workspace.sidebar.options.menus.constraints :refer [constraint-attrs constraints-menu]]
+   [app.main.ui.workspace.sidebar.options.menus.exports :refer [exports-menu* exports-attrs]]
    [app.main.ui.workspace.sidebar.options.menus.fill :as fill]
    [app.main.ui.workspace.sidebar.options.menus.grid-cell :as grid-cell]
    [app.main.ui.workspace.sidebar.options.menus.layer :refer [layer-attrs layer-menu]]
@@ -23,17 +24,11 @@
    [rumext.v2 :as mf]))
 
 (mf/defc options*
-  [{:keys [shape] :as props}]
-  (let [ids
-        (mf/with-memo [shape]
-          [(dm/get-prop shape :id)])
-
-        shapes
-        (mf/with-memo [shape]
-          [shape])
-
-        type
-        (dm/get-prop shape :type)
+  [{:keys [shape file-id page-id]}]
+  (let [id     (dm/get-prop shape :id)
+        type   (dm/get-prop shape :type)
+        ids    (mf/with-memo [id] [id])
+        shapes (mf/with-memo [shape] [shape])
 
         measure-values
         (select-keys shape measure-attrs)
@@ -131,4 +126,11 @@
      [:& blur-menu {:ids ids
                     :values (select-keys shape [:blur])}]
      [:& svg-attrs-menu {:ids ids
-                         :values (select-keys shape [:svg-attrs])}]]))
+                         :values (select-keys shape [:svg-attrs])}]
+     [:> exports-menu* {:type type
+                        :ids ids
+                        :shapes shapes
+                        :values (select-keys shape exports-attrs)
+                        :page-id page-id
+                        :file-id file-id}]]))
+