+++ frontend/src/app/main/ui/workspace/sidebar/options/shapes/text.cljs
@@ -16,6 +16,7 @@
    [app.main.ui.workspace.sidebar.options.menus.blur :refer [blur-menu]]
    [app.main.ui.workspace.sidebar.options.menus.color-selection :refer [color-selection-menu*]]
    [app.main.ui.workspace.sidebar.options.menus.constraints :refer [constraint-attrs constraints-menu]]
+   [app.main.ui.workspace.sidebar.options.menus.exports :refer [exports-menu* exports-attrs]]
    [app.main.ui.workspace.sidebar.options.menus.fill :as fill]
    [app.main.ui.workspace.sidebar.options.menus.grid-cell :as grid-cell]
    [app.main.ui.workspace.sidebar.options.menus.layer :refer [layer-attrs layer-menu]]
@@ -28,7 +29,7 @@
    [rumext.v2 :as mf]))
 
 (mf/defc options*
-  [{:keys [shape file-id libraries] :as props}]
+  [{:keys [shape libraries file-id page-id]}]
   (let [id     (dm/get-prop shape :id)
         type   (dm/get-prop shape :type)
         ids    (mf/with-memo [id] [id])
@@ -186,4 +187,12 @@
 
      [:& blur-menu
       {:ids ids
-       :values (select-keys shape [:blur])}]]))
+       :values (select-keys shape [:blur])}]
+
+     [:> exports-menu* {:type type
+                        :ids ids
+                        :shapes shapes
+                        :values (select-keys shape exports-attrs)
+                        :page-id page-id
+                        :file-id file-id}]]))
+